Fix issues with ^^ color escape failing
Issue #412

q_shared.h - Added Q_IsColorEscape() and changed Q_IsColorString() to avoid interpreting ^\n as a color string. This was leading to problems in the console because the \n was being stripped from the buffer.
cg_draw - Fixes the chat area at the bottom of the screen
cl_console - Fixes the console's scrollback display
cl_scrn - Fixes the console's input display
ui_main - Fixes the editfield in the chat dialog

The bug is a little more extensive than this. Pretty much any case where it was doing if (Q_IsColorString()) needs to have a preceeding Q_IsColorEscape(). I changed the cases visually affected, and then a couple logical strlen()s. I believe this is at least stable and fixes all the common spots where the bug's effect was visible, but not every spot.