Fixes issue #412
The first problem described is (when connected to a server) typing into the console: "^<Return>more<Return>" where <Return> means pressing the return key, the scrollback would show:

]Bob: ]more
Bob: more

Instead of:

]^
Bob: ^
]more
Bob: more

This was caused by Q_IsColorString() being inaccurate and identifying "^\n" as a color string, resulting in the \n never ending up in the buffer.

The change to Q_IsColorString() fixes this by explicitly whitelisting the valid range of chars that map to g_color_table and the COLOR_NULL char.
    So definitively:
        *0123456789:;<=>?ABCDEFGHIJKLMNO

    Along with lowercase a-o for convenience.