cgame: add bp command, refs #233 #1198
This adds message area on the top of the screen.

Can be used to print banners or other messages in a less
distractive way.

This does not implement the banner rotator,
for such feature check out the lua script:
https://github.com/etlegacy/etlegacy-lua_scripts#banners

Client cvars to control it:
cg_drawBannerPrint - toggle the message area
cg_bannerPrintTime - duration the message is displayed